+++ eth_defi/middleware.py
@@ -15,13 +15,10 @@ import logging
 import time
 from http.client import RemoteDisconnected
 from pprint import pformat
-from typing import (Any, Callable, Collection, Counter, Optional, Tuple, Type,
-                    TypeAlias, Union)
+from typing import Any, Callable, Collection, Counter, Optional, Tuple, Type, TypeAlias, Union
 
 from eth_utils.toolz import assoc
-from requests.exceptions import (ChunkedEncodingError, ConnectionError,
-                                 ContentDecodingError, HTTPError, Timeout,
-                                 TooManyRedirects)
+from requests.exceptions import ChunkedEncodingError, ConnectionError, ContentDecodingError, HTTPError, Timeout, TooManyRedirects
 from web3 import Web3
 from web3._utils.transactions import get_buffered_gas_estimate
 from web3.exceptions import BlockNotFound
@@ -29,8 +26,7 @@ from web3.middleware import Middleware
 from web3.types import RPCEndpoint, RPCResponse
 
 from eth_defi.compat import WEB3_PY_V7, check_if_retry_on_failure_compat
-from eth_defi.compat import \
-    exception_retry_middleware as compat_exception_retry_middleware
+from eth_defi.compat import exception_retry_middleware as compat_exception_retry_middleware
 from eth_defi.tx import get_tx_broadcast_data
 
 logger = logging.getLogger(__name__)